requirements.txt
streamlit 
streamlit-chat
# streamlit-extras
python-dotenv
aiohttp #This is for telegram
requests
asyncio
gradio_client
deep_translator
groq
pillow